+++ .gitignore
@@ -1,0 +1,5 @@
+*.class
+*.project
+.settings/
+bin/
+.DS_Store
