+++ README.md
@@ -1,0 +1,8 @@
+# McWiki
+
+"For the noob in all of us".
+
+This plugin aims to present information that can only be accessed through the minecraft wiki (externally) in game (internally). 
+
+Spigotmc link: https://www.spigotmc.org/resources/mcwiki.35039/
+Github link: https://github.com/skylerdev/mcwiki
